requirements.txt
Flask==1.1.1
gunicorn==19.9.0
MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy>=1.17.2
scipy>=0.15.1
scikit-learn==0.22.2
matplotlib>=1.4.3
pandas>=0.19
jsonify
joblib
requests
tensorflow-cpu
